audit: P4++++++++++++++++++++ edison-advance 九轮审计修复
- TransportBar::resized() 移除未使用的 selInfoArea 变量
- RegionOverlay 析构时清理 onRegionsChanged 回调避免悬挂
- SpectrogramComponent::renderViewport 同步 m_viewOffset 钳位值
- SpectrogramComponent::setPlaybackPosition 自动滚动上限钳位 (clamp)
- DragExport 临时文件定期清理（最多保留 20 个）

+++ Source/LoopOverlay.cpp
@@ -25,8 +25,8 @@ void LoopOverlay::paint(juce::Graphics& g)
     loopStart = juce::jlimit(0.0, totalDur, loopStart);
     loopEnd   = juce::jlimit(0.0, totalDur, loopEnd);
 
-    float x1 = m_waveformThumbnail.timeToX(static_cast<float>(loopStart));
-    float x2 = m_waveformThumbnail.timeToX(static_cast<float>(loopEnd));
+    float x1 = m_waveformThumbnail.timeToX(loopStart);
+    float x2 = m_waveformThumbnail.timeToX(loopEnd);
 
     auto loopRect = juce::Rectangle<float>(x1, 0.0f, x2 - x1, area.getHeight());
